+++ webapp/javascript/redux/reducers/continuous/singleView.thunks.ts
@@ -1,0 +1,41 @@
+import { renderSingle, RenderOutput } from '@webapp/services/render';
+import { RequestAbortedError } from '@webapp/services/base';
+import { addNotification } from '../notifications';
+import { createAsyncThunk } from '../../async-thunk';
+import { ContinuousState } from './state';
+
+let singleViewAbortController: AbortController | undefined;
+
+export const fetchSingleView = createAsyncThunk<
+  RenderOutput,
+  null,
+  { state: { continuous: ContinuousState } }
+>('continuous/singleView', async (_, thunkAPI) => {
+  if (singleViewAbortController) {
+    singleViewAbortController.abort();
+  }
+
+  singleViewAbortController = new AbortController();
+  thunkAPI.signal = singleViewAbortController.signal;
+
+  const state = thunkAPI.getState();
+  const res = await renderSingle(state.continuous, singleViewAbortController);
+
+  if (res.isOk) {
+    return Promise.resolve(res.value);
+  }
+
+  if (res.isErr && res.error instanceof RequestAbortedError) {
+    return Promise.reject(res.error);
+  }
+
+  thunkAPI.dispatch(
+    addNotification({
+      type: 'danger',
+      title: 'Failed to load single view data',
+      message: res.error.message,
+    })
+  );
+
+  return Promise.reject(res.error);
+});
